requirements.txt
langgraph>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
openai>=1.0.0
chromadb>=1.0.0
tiktoken>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
